// barq-system-frontend/src/sdk/types.gen.ts
export type AlliancesClientsEntity = {
  id: number;
  media_id: number;
  media: MediaEntity;
  countries: Array<CountryEntity>;
  industries: Array<IndustriesEntity>;
  created_at: string;
  updated_at: string;
  deleted_at: string;
};

export type AlliancesHeadEntity = {
  id: number;
  title: string;
  sub_title: string;
  alliances_head_id_alliances_head_translations: Array<AlliancesHeadTranslationsEntity>;
  created_at: string;
  updated_at: string;
};

export type AlliancesHeadTranslationsEntity = {
  id: number;
  alliances_head_id: number;
  alliances_head_translation_alliances_head?: AlliancesHeadEntity | null;
  title: string;
  sub_title: string;
  language: 'ar' | 'en';
};

export type AlliancesVendorsEntity = {
  id: number;
  media_id: number;
  media: MediaEntity;
  countries: Array<CountryEntity>;
  solutions: Array<SolutionsEntity>;
  created_at: string;
  updated_at: string;
  deleted_at: string;
};
// ...
export type CityEntity = {
  id: number;
  name: string;
  city_id_city_translations: Array<CityTranslationsEntity>;
  country_id: number;
  country: CountryEntity;
  career_open_position_city: Array<CareerOpenPositionEntity>;
  created_at: string;
  updated_at: string;
  deleted_at: string;
};
// ...
export type CountryEntity = {
  id: number;
  name: string;
  country_id_country_translations: Array<CountryTranslationsEntity>;
  country_id_alliances_vendors: Array<AlliancesVendorsEntity>;
  country_id_alliances_clients: Array<AlliancesClientsEntity>;
  country_id_cities: Array<CityEntity>;
  country_id_success_story_case_studies: Array<SuccessStoryCaseStudiesEntity>;
  created_at: string;
  updated_at: string;
  deleted_at: string;
};

export type CountryTranslationsEntity = {
  id: number;
  country_id: number;
  country_translation_country?: CountryEntity | null;
  name: string;
  language: 'ar' | 'en';
};
// ...
export type type = 'soc' | 'cybersecurity' | 'grc';
// ...
export type IndustriesEntity = {
  id: number;
  name: string;
  industries_id_industries_translations: Array<IndustriesTranslationsEntity>;
  industries_id_alliances_clients: Array<AlliancesClientsEntity>;
  industries_id_success_story_case_studies: Array<SuccessStoryCaseStudiesEntity>;
  created_at: string;
  updated_at: string;
  deleted_at: string;
};

export type IndustriesTranslationsEntity = {
  id: number;
  industries_id: number;
  industries_translation_industries?: IndustriesEntity | null;
  name: string;
  language: 'ar' | 'en';
};
// ...
export type id = 'ASC' | 'DESC';

export type title = 'ASC' | 'DESC';

export type description = 'ASC' | 'DESC';

export type image_id = 'ASC' | 'DESC';

export type created_at = 'ASC' | 'DESC';

export type updated_at = 'ASC' | 'DESC';
// ...
export type sub_title = 'ASC' | 'DESC';

export type media_id = 'ASC' | 'DESC';
// ...
export type name = 'asc' | 'desc';
// ...
export type deleted_at = 'asc' | 'desc';
// ...
export type number = 'asc' | 'desc';
// ...
export type SolutionsEntity = {
  id: number;
  name: string;
  solutions_id_solutions_translations: Array<SolutionsTranslationsEntity>;
  solutions_id_alliances_vendors: Array<AlliancesVendorsEntity>;
  created_at: string;
  updated_at: string;
  deleted_at: string;
};

export type SolutionsTranslationsEntity = {
  id: number;
  solutions_id: number;
  solutions_translation_solutions?: SolutionsEntity | null;
  name: string;
  language: 'ar' | 'en';
};

export type SuccessStoryCaseStudiesEntity = {
  id: number;
  title: string;
  description: string;
  long_description: string;
  cta_button_text: string;
  featured: boolean;
  is_featured: boolean;
  date: string;
  image_id: number;
  home_image_id: number;
  read_time: number;
  image: MediaEntity;
  home_image: MediaEntity;
  industries_id: number;
  industries: IndustriesEntity;
  country_id: number;
  country: CountryEntity;
  success_story_case_studies_id_success_story_case_studies_translations: Array<SuccessStoryCaseStudiesTranslationsEntity>;
  created_at: string;
  updated_at: string;
  deleted_at: string;
};
// ...
export type CountryControllerReadResponse = {
  data?: Array<CountryEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};
// ...
export type IndustriesControllerReadResponse = {
  data?: Array<IndustriesEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};
// ...
export type SolutionsControllerReadResponse = {
  data?: Array<SolutionsEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};
// ...
export type AlliancesVendorsControllerReadResponse = {
  data?: Array<AlliancesVendorsEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};
// ...
export type AlliancesClientsControllerReadResponse = {
  data?: Array<AlliancesClientsEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};
// ...
export type AlliancesHeadControllerReadResponse = {
  data?: Array<AlliancesHeadEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};